some changes on the DHW multiplier a areabasedflowrate default is added for building that has 2 different ventilation systems these are dealt wihtin load and occupancy script for each zone

+++ CoreFiles/GeneralFunctions.py
@@ -103,7 +103,7 @@ def SaveCase(MainPath,SepThreads,CaseName,nbBuild):
         except:
             pass
 
-def CreateSimDir(CurrentPath,CaseName,SepThreads,nbBuild,idx,LogFile):
+def CreateSimDir(CurrentPath,CaseName,SepThreads,nbBuild,idx,LogFile,Refrech = False):
     if not os.path.exists(os.path.join(os.path.dirname(os.path.dirname(CurrentPath)),'MUBES_SimResults')):
         os.mkdir(os.path.join(os.path.dirname(os.path.dirname(CurrentPath)),'MUBES_SimResults'))
     SimDir = os.path.normcase(
@@ -111,7 +111,7 @@ def CreateSimDir(CurrentPath,CaseName,SepThreads,nbBuild,idx,LogFile):
     LogFileName = CaseName + '_Logs.log'
     if not os.path.exists(SimDir):
         os.mkdir(SimDir)
-    elif idx == 0:
+    elif idx == 0 and Refrech:
         shutil.rmtree(SimDir)
         os.mkdir(SimDir)
     if SepThreads:
